// src/lib/types/enums/action-enum.ts
export const enum ActionEnum {
	Dynamic = 'dynamic',
	CreateObjects = 'create',
	AddVersions = 'addversion',
	UpdateObjects = 'update',
	MoveObjects = 'move',
	CopyObjects = 'copy',
	PurgeVersions = 'purge',
	DeleteObjects = 'delete',
	DeleteRenditions = 'deleterendition',
	AssignCrossReference = 'assignxref'
}