// src/lib/models/import/import-csv-job-model.ts
import type { TDateISO } from '$lib/types/date-time-type';
import type { JobModel } from '$lib/models/job-model';
import type { TargetPathMeaningEnum } from '$lib/types/enums/target-path-meaning-enum';
import type { ContainerCollisionOptionsEnum } from '$lib/types/enums/container-collision-options-enum';
import type { DocOptionAddVersionEnum } from '$lib/types/enums/doc-add-newer-enum';
import type { ContainerTypeEnum } from '$lib/types/enums/container-type-enum';
import type { RootFileLocEnum } from '$lib/types/enums/root-file-loc-enum';
import type { DocCollisionOptionsEnum } from '$lib/types/enums/doc-collision-options-enum';
import type { ActionEnum } from '$lib/types/enums/action-enum';
import type { AttrSourceTypeEnum } from '$lib/types/enums/attr-source-type-enum';

export interface ImportCsvJobModel extends JobModel {
	action: ActionEnum;
	actionMetadata?: boolean;
	actionRename?: boolean;
	archiveCsvFiles?: boolean;
	archiveImportedFiles?: boolean;
	attrSourceType: AttrSourceTypeEnum;
	autoCreateFolders?: boolean;
	checkRootPath?: boolean;
	colNames?: string[];
	containerOption: ContainerCollisionOptionsEnum;
	containerOptionUpdateMd?: boolean;
	containerType: ContainerTypeEnum;
	csvArchivePath: string;
	defaultFileExtension: string;
	docOption: DocCollisionOptionsEnum;
	docOptionAddVersion?: boolean;
	docOptionAddVersionIfNewer?: boolean;
	docOptionAddVersionRadioVal?: DocOptionAddVersionEnum;
	docOptionUpdateMd?: boolean;
	fileArchivePath: string;
	forcePermInherit?: boolean;
	includeRoot?: boolean;
	inheritParent?: boolean;
	mdCats?: boolean;
	mdDapi?: boolean;
	newProfileId?: number | null;
	profileId?: number;
	rmClassInherit?: boolean;
	rootFileLoc: RootFileLocEnum;
	rootSourcePath: string;
	rootTargetPath: string;
	sourceOption?: string;
	specific_at_time_1?: TDateISO | null;
	targetContainer?: string;
	targetPathMeaning: TargetPathMeaningEnum;
	useFileCreateDate?: boolean;
	useFileModifyDate?: boolean;
}